ISSUE-5421 emailAddressContactEvent RabbitMQEventBus should not run when Redis event bus
When Redis event bus mode:
`emailAddressContactEvent-workQueue-org.apache.james.events.TmailGroupRegistrationHandler$GroupRegistrationHandlerGroup` was created by Redis event bus (of course)
`emailAddressContactEvent-workQueue-org.apache.james.events.GroupRegistrationHandler$GroupRegistrationHandlerGroup` was created by RabbitMQEventBus which should not.
